+++ src/Ampersand/Prototype/GenAngularFrontend.hs
@@ -34,8 +34,8 @@ crudsToString x = T.pack $ zipWith (curry f) [crudC x, crudR x, crudU x, crudD x
 
 genComponent :: (HasRunner env, HasDirPrototype env) => FSpec -> FEInterface -> RIO env ()
 genComponent fspec ifc = do
-  genComponentTs fspec ifc
-  genComponentView fspec ifc
+  unless (isApi ifc) $ genComponentTs fspec ifc
+  unless (isApi ifc) $ genComponentView fspec ifc
   genComponentInterface fspec ifc
   logInfo . display $ "Generated files for " <> ifcNamePascal ifc <> "Component"
 
@@ -107,7 +107,9 @@ genSingleFileFromTemplate fSpec feSpec templateFilePath targetFilePath = do
         renderTemplate Nothing template $
           setAttribute "contextName" (fsName fSpec)
             . setAttribute "ampersandVersionStr" (longVersion appVersion)
-            . setAttribute "ifcs" (interfaces feSpec)
+            . setAttribute "ifcs" (interfaces feSpec) -- all interfaces
+            . setAttribute "uis" (filter (not . isApi) $ interfaces feSpec) -- only the interfaces that need UI
+            . setAttribute "apis" (filter isApi $ interfaces feSpec) -- only the interfaces that have API (no UI)
             . setAttribute "concepts" (concepts feSpec)
             . setAttribute "views" (views feSpec)
             . setAttribute "verbose" (loglevel' == LevelDebug)